Move DocumentationPage class into Models\Pages namespace
