+++ packages/framework/resources/views/pages/documentation-search.blade.php
@@ -1,13 +1,19 @@
 @php
-	$page = new \Hyde\Framework\Models\DocumentationPage([], '', 'Search', 'search');
-	$title = 'Search';
-	$currentPage = $page->getCurrentPagePath();
-	$markdown = '';
+    $page = new \Hyde\Framework\Models\Pages\DocumentationPage([], '', 'Search', 'search');
+    $title = 'Search';
+    $currentPage = $page->getCurrentPagePath();
+    $markdown = '';
 @endphp
 
 @extends('hyde::layouts.docs')
 @section('content')
-	<h1>Search the documentation site</h1>
-	<style>#searchMenuButton{display:none!important;}#search-results{max-height:unset!important;}</style>
-	@include('hyde::components.docs.search-input')
+    <h1>Search the documentation site</h1>
+    <style>#searchMenuButton {
+            display: none !important;
+        }
+
+        #search-results {
+            max-height: unset !important;
+        }</style>
+    @include('hyde::components.docs.search-input')
 @endsection
