+++ packages/framework/src/Actions/GeneratesDocumentationSearchIndexFile.php
@@ -5,7 +5,7 @@ namespace Hyde\Framework\Actions;
 use Hyde\Framework\Concerns\InteractsWithDirectories;
 use Hyde\Framework\Contracts\ActionContract;
 use Hyde\Framework\Hyde;
-use Hyde\Framework\Models\DocumentationPage;
+use Hyde\Framework\Models\Pages\DocumentationPage;
 use Hyde\Framework\Models\Parsers\DocumentationPageParser;
 use Hyde\Framework\Services\CollectionService;
 use Illuminate\Support\Collection;
