+++ packages/framework/src/Commands/HydeMakePageCommand.php
@@ -5,7 +5,7 @@ namespace Hyde\Framework\Commands;
 use Hyde\Framework\Actions\CreatesNewPageSourceFile;
 use Hyde\Framework\Exceptions\UnsupportedPageTypeException;
 use Hyde\Framework\Models\Pages\BladePage;
-use Hyde\Framework\Models\DocumentationPage;
+use Hyde\Framework\Models\Pages\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
 use LaravelZero\Framework\Commands\Command;
