+++ packages/framework/src/Concerns/Internal/FileHelpers.php
@@ -2,7 +2,7 @@
 
 namespace Hyde\Framework\Concerns\Internal;
 
-use Hyde\Framework\Models\DocumentationPage;
+use Hyde\Framework\Models\Pages\DocumentationPage;
 
 /**
  * Offloads file helper methods for the Hyde Facade.
