+++ packages/framework/src/Concerns/Internal/FluentPathHelpers.php
@@ -3,7 +3,7 @@
 namespace Hyde\Framework\Concerns\Internal;
 
 use Hyde\Framework\Models\Pages\BladePage;
-use Hyde\Framework\Models\DocumentationPage;
+use Hyde\Framework\Models\Pages\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
 use Hyde\Framework\Models\Pages\MarkdownPost;
 use Hyde\Framework\Services\DiscoveryService;
