+++ packages/framework/src/Services/DiscoveryService.php
@@ -4,7 +4,7 @@ namespace Hyde\Framework\Services;
 
 use Hyde\Framework\Contracts\AbstractPage;
 use Hyde\Framework\Models\Pages\BladePage;
-use Hyde\Framework\Models\DocumentationPage;
+use Hyde\Framework\Models\Pages\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
 use Hyde\Framework\Models\Pages\MarkdownPost;
