+++ packages/framework/tests/Feature/HydeDocsIndexPathTest.php
@@ -3,7 +3,7 @@
 namespace Hyde\Framework\Testing\Feature;
 
 use Hyde\Framework\Hyde;
-use Hyde\Framework\Models\DocumentationPage;
+use Hyde\Framework\Models\Pages\DocumentationPage;
 use Hyde\Testing\TestCase;
 
 class HydeDocsIndexPathTest extends TestCase
