+++ packages/framework/tests/Feature/Services/HydeSmartDocsTest.php
@@ -4,7 +4,7 @@ namespace Hyde\Framework\Testing\Feature\Services;
 
 use Hyde\Framework\Actions\MarkdownConverter;
 use Hyde\Framework\Hyde;
-use Hyde\Framework\Models\DocumentationPage;
+use Hyde\Framework\Models\Pages\DocumentationPage;
 use Hyde\Framework\Services\HydeSmartDocs;
 use Hyde\Testing\TestCase;
