+++ packages/realtime-compiler/src/Routing/PageRouter.php
@@ -6,7 +6,7 @@ use Desilva\Microserve\Request;
 use Desilva\Microserve\Response;
 use Hyde\Framework\Hyde;
 use Hyde\Framework\Models\Pages\BladePage;
-use Hyde\Framework\Models\DocumentationPage;
+use Hyde\Framework\Models\Pages\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
 use Hyde\Framework\Models\Pages\MarkdownPost;
 use Hyde\RealtimeCompiler\Actions\Compiler;
